requirements.txt
flask
PyPDF2
werkzeug
python-docx